+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Fix O(n) performance issue in `AtomicThreadPool` queue by using two-stack linked list
+
 ## [0.10.0] - 2025-11-09
 
 - Add `AtomicThreadPool#active_count`